Keep the Data members around while they are being used.
What we do here is to create a global map between a
Data pointer, and its shared_ptr equivalent.  When we
create a new Data ptr via rmw_context_impl_s, we add
it to the map.  When we delete a Data ptr via rmw_context_impl_s,
we remove it from the map.  When we get a graph callback,
we look to see if the pointer is in the map; if so, we
get the shared_ptr.  Because of this, we are guaranteed
that the Data pointer will not be removed while we are
using it.

Signed-off-by: Chris Lalancette <clalancette@gmail.com>