Don't block in rmw_init checking for the router.
It is much cleaner to just have rmw_init return without
looping, so we switch to only checking for the router
without waiting.  However, if we don't find a router,
warn the user.

It still works to start a publisher and subscription, then
start the router; they'll connect at that point.

Signed-off-by: Chris Lalancette <clalancette@gmail.com>