Updates to compile against Humble.
There are 5 main changes in here:
1.  Humble does not support type hashes.  Rather than change
all of the liveliness token handling (which will make backports
in the future harder), we just remove the code that computes it,
and instead set it to a constant string.
2.  The definition of RMW_GID_STORAGE_SIZE in Humble is wonky,
and set to 24 for some ancient historical reasons.  We don't
actually want to do that, so just hard-code 16.
3.  Some of the events in modern ROS 2 aren't supported in Humble,
so remove the code for handling them.
4.  There are a few APIs dealing with dynamic messages that aren't
supported in Humble.
5.  There is at least one QOS policy (best available) which doesn't
exist in Humble.

With all of these changes in, I'm able to compile and run basic
programs against Humble.

Signed-off-by: Chris Lalancette <clalancette@gmail.com>